Fixed propagation of weights throughout learning.
Weights should be adjusted by differentials after backwards run. Original
implementation however was using already adjusted weights for calculations
withing same backward run. This was fixed.

+++ src/main/util/cpuDebugHelpers.h
@@ -8,7 +8,7 @@
 #ifndef CPUDEBUGHELPERS_H
 #define	CPUDEBUGHELPERS_H
 
-void dumpHostArray(char flag, double *array, int size);
+void dumpHostArray(char flag, float *array, int size);
 
 #endif	/* CPUDEBUGHELPERS_H */
